test: move some fullscreen tests, cleanup helpers
